Configurable actions & CLI Improvements (#40)
* Fixed printing bug, fixed some cli warnings, added config to the cli, added printing of timers and config. Not done yet

* CLI stuff

* Fixed printing issue and config saving issue

* Fixed printing bug, fixed some cli warnings, added config to the cli, added printing of timers and config. Not done yet

* CLI stuff

* Fixed printing issue and config saving issue

* Ready for review

* Addressed issues from pull request

* formatting of settings.c

* Update header of cli files, to include source

* Code review with Nemanja

+++ boards/cats_rev1Pro/src/util/actions.h
@@ -45,4 +45,4 @@ typedef enum {
 } action_function_list_e;
 
 /* TODO - don't export this anymore after the flash is working */
-bool set_recorder_state(int32_t state);
+bool set_recorder_state(int16_t state);
